move iso8601 to maintenance script for dsc compat
